+++ examples/basic/src/main/java/net/zerobuilder/examples/inheritance/Planet.java
@@ -2,12 +2,14 @@ package net.zerobuilder.examples.inheritance;
 
 import net.zerobuilder.Build;
 
+import java.math.BigInteger;
+
 @Build(toBuilder = true)
 final class Planet extends CelestialBody {
   final int numberOfMoons;
   private final boolean habitable;
 
-  Planet(long mass, int numberOfMoons, boolean habitable) {
+  Planet(BigInteger mass, int numberOfMoons, boolean habitable) {
     super(mass);
     this.numberOfMoons = numberOfMoons;
     this.habitable = habitable;
